Move logic inside of MVVM in theme mode and add close when scrolling
