+++ src/stories/components/menu-navigation/menu-theme/menu-theme.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
-import { useThemeContext } from '../../../../core/context/ThemeContext';
 import { ThreeDots } from '../../svg/three-dots';
 import styled from '@emotion/styled';
 import { ThemeType } from '../../../../core/enums/theme.enum';
 import MenuItemTheme from './menu-item-theme';
+import useMenuThemeMVVM from './menu-theme.mvvm';
 
 interface Props {
   themeMode: ThemeType;
@@ -21,15 +21,7 @@ export const MenuPaperStyle = (isLight: boolean) => ({
 });
 
 const MenuTheme = ({ themeMode, toggleTheme }: Props) => {
-  const { isLight } = useThemeContext();
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
+  const { isLight, anchorEl, handleClick, open, handleClose } = useMenuThemeMVVM();
 
   return (
     <>
